Module V1.EZ_SUBST

$

"

let s = EZ_SUBST.string ~sep:'!' ~paren:(function _ s -> String.uppercase s) () "$(abc) $(def)"

let s = EZ_SUBST.string ~sym:true ~sep:'%' ~brace:(fun ctxt_ s -> ctxt ^ " " ^ s) "Hello" "%John% %Sandy%"

type 'context t = 'context -> string -> string

The type for functions performing the translation from ident to its replacement. 'context is some information, that is from the initial call to the substitution.

exception UnclosedExpression of string

The only exception that may be raised by substitutions: it indicates that the end of the expression could not be found.

val string : ?⁠sep:char -> ?⁠sym:bool -> ?⁠escape:bool Stdlib.ref -> ?⁠brace:'context t -> ?⁠paren:'context t -> ?⁠bracket:'context t -> ?⁠var:'context t -> 'context -> string -> string

string f context s performs substitutions on s following f, passing the context context to f for every expression, returning the result as a string.

val buffer : ?⁠sep:char -> ?⁠sym:bool -> ?⁠escape:bool Stdlib.ref -> ?⁠brace:'context t -> ?⁠paren:'context t -> ?⁠bracket:'context t -> ?⁠var:'context t -> Stdlib.Buffer.t -> 'context -> string -> unit

buffer f b context s performs substitutions on s following f, passing the context context to f for every expression, returning the result by appending it to the buffer b.

Note that modifications are not atomic, so if an exception is raised during the substitution, the buffer might have been modified.